iSeeFit-py/iseefit/requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 数据验证
pydantic[email]==2.5.0
email-validator==2.1.0

# 图片处理
Pillow>=9.0.0

# 环境变量
python-dotenv==1.0.0

# 日志
loguru==0.7.2

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
